feedbackFinalTimes

+++ widgets/cut.cpp
@@ -896,7 +896,7 @@ void Cut::findNextMark(int mark)
 		}
 
 		int index = std::distance(std::begin(markArr), it);
-		ui->paintStartDoubleSpinBox->setValue(std::max(0., double(index) / edfFil.getFreq() - 0.5));
+		ui->paintStartDoubleSpinBox->setValue(std::max(0., index / edfFil.getFreq() - 0.5));
 //		ui->leftLimitSpinBox->setValue(index);
 		showDerivatives();
 		paint();
@@ -938,7 +938,7 @@ void Cut::findPrevMark(double mark)
 		}
 
 		int index = std::distance(beg, it);
-		ui->paintStartDoubleSpinBox->setValue(std::max(0., double(index) / edfFil.getFreq() - 0.5));
+		ui->paintStartDoubleSpinBox->setValue(std::max(0., index / edfFil.getFreq() - 0.5));
 //		ui->leftLimitSpinBox->setValue(index);
 		showDerivatives();
 		paint();
